change whole network into float16 precision

+++ OCTMultiSurfaces/network/OCTDataSet.py
@@ -9,12 +9,12 @@ class OCTDataSet(data.Dataset):
     def __init__(self, imagesPath, labelPath, IDPath, transform=None, device=None, sigma=20.0):
         self.m_device = device
         self.m_sigma = sigma
-        images = torch.from_numpy(np.load(imagesPath).astype(np.float32)).to(self.m_device, dtype=torch.float)  # slice, H, W
+        images = torch.from_numpy(np.load(imagesPath).astype(np.float32)).to(self.m_device, dtype=torch.float16)  # slice, H, W
         # normalize images for each slice
         std,mean = torch.std_mean(images, dim=(1,2))
         self.m_images = TF.Normalize(mean, std)(images)
 
-        self.m_labels = torch.from_numpy(np.load(labelPath).astype(np.float32)).to(self.m_device, dtype=torch.float)  # slice, num_surface, W
+        self.m_labels = torch.from_numpy(np.load(labelPath).astype(np.float32)).to(self.m_device, dtype=torch.float16)  # slice, num_surface, W
         with open(IDPath) as f:
             self.m_IDs = json.load(f)
         self.m_transform = transform
@@ -43,7 +43,7 @@ class OCTDataSet(data.Dataset):
         Num_Surface, W = Mu.shape
         Mu = Mu.unsqueeze(dim=-2)
         Mu = Mu.expand((Num_Surface,H,W))
-        X = torch.arange(start=0.0, end=H, step=1.0).view((H,1)).to(self.m_device, dtype=torch.float)
+        X = torch.arange(start=0.0, end=H, step=1.0).view((H,1)).to(self.m_device, dtype=torch.float16)
         X = X.expand((Num_Surface,H,W))
         pi = math.pi
         G  = torch.exp(-(X-Mu)**2/(2*sigma*sigma))/(sigma*math.sqrt(2*pi))
